Teal: make luarocks.fs not dependent on luarocks.util
Break circular dependency detected by Cyan
